set MTU to 1500

+++ src/os_frame.rs
@@ -3,7 +3,7 @@ use crate::MTU;
 /// Representation of a network packet that can be interpreted by specific OSs
 /// Packets of this kind can either be in raw IP or null/loopback form
 pub struct OsFrame {
-    pub frame: [u8; MTU as usize],
+    pub frame: [u8; MTU],
     pub actual_bytes: usize,
 }
 
@@ -13,7 +13,7 @@ impl OsFrame {
 
     pub fn new() -> Self {
         Self {
-            frame: [0; MTU as usize],
+            frame: [0; MTU],
             actual_bytes: 0,
         }
     }
